Create Suspend Users from CSV
This will suspend all of the users in a CSV list, by email, and then generate a confirmation report that states who all it found and suspended. 
This is often because, as things are, sometimes we get lists of people that may not even have Canvas accounts, but others believe they do, so it's good to have a final report of who all it actually found and suspended. Thank you to all those that helped generate this!